Move CommandColumn into FreeMode
