--- src/components/CommandColumn/index.ts
@@ -1,3 +1,0 @@
-import CommandColumn from "./CommandColumn"
-
-export default CommandColumn
