+++ src/components/FreeMode/CommandColumn.scss
