+++ src/components/FreeMode/CommandColumn.tsx
